Fix for JIRA-16889: -Attempt to remove a directory only if it is empty
svn path=/trunk/; revision=47696


Former-commit-id: 71842e81bcc88f859d65a4fc4f855a9c79a17046

+++ v3/cluster/cli/src/main/java/com/sun/enterprise/admin/cli/cluster/SSHCommandsBase.java
@@ -219,10 +219,15 @@ public abstract class SSHCommandsBase extends CLICommand {
                 //using force deletes all files on remote host
                 if(force) {
                     logger.fine("Force removing directory " + f1);
-                    sftpClient.rmdir(f1); 
+                    if (isRemoteDirectoryEmpty(sftpClient, f1)) {
+                        sftpClient.rmdir(f1); 
+                    }
                 } else {                    
-                    if (dasFiles.contains(f1))
-                        sftpClient.rmdir(f1);
+                    if (dasFiles.contains(f1)) {
+                        if (isRemoteDirectoryEmpty(sftpClient, f1)) {
+                            sftpClient.rmdir(f1); 
+                        }
+                    }
                 }
             } else {
                 String f2 = dir+"/"+directoryEntry.filename;
@@ -235,6 +240,21 @@ public abstract class SSHCommandsBase extends CLICommand {
                 }
             }
         }
+    }
+
+    /**
+     * Method to check if specified remote directory contains files
+     * 
+     * @param sftp SFTP client handle
+     * @param file path to remote directory
+     * @return true if empty, false otherwise
+     * @throws IOException
+     */
+    private boolean isRemoteDirectoryEmpty(SFTPClient sftp, String file) throws IOException {
+        List<SFTPv3DirectoryEntry> l = (List<SFTPv3DirectoryEntry>)sftp.ls(file);        
+        if (l.size() > 2)
+            return false;
+        return true;
     }
     
     /**
